+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.43.0 (2024-mm-dd)
 
 - **ENH: Add prune and select keys from dict functions: `misc.select_dict` and `misc.prune_dict`**
+- **ENH: Handle dict in `misc.remove_empty_values` and allow the user to add other values considered as empty**
 - FIX: Select arguments from `**kwargs` for functions with finished number of arguments
 
 ## 1.42.3 (2024-11-04)